added private share api
